+++ theme/template/posts/sponsor.php
@@ -4,7 +4,7 @@
       <div class="col-12">
         <section class="subpage__hero">
           <h1 class="subpage__hero-title">
-            <?= $post->post_title ?>
+            Sponsors
           </h1>
         </section>
       </div>
@@ -15,14 +15,20 @@
 <section class="wrapper">
   <div class="container">
     <div class="row">
-      <div class="col-12 col-sm-10 offset-sm-1">
-        <main class="article">
-          <?= $post->post_content ?>
-        </main>
+      <div class="col-12 col-sm-4">
+        <img class="img-fluid" src="<?= get_template_directory_uri() ?>/assets/images/placeholder.png">
+      </div>
 
-        <footer class="text-center">
-          <?php include get_template_directory() . '/_includes/social.php'; ?>
-        </footer>
+      <div class="col-12 col-sm-8">
+        <main class="article">
+          <?= $post->post_content; ?>
+
+          <hr>
+
+          <a href="/<?= SPONSORS_PAGE; ?>/">
+            一覧に戻る
+          </a>
+        </main>
       </div>
     </div>
   </div>